+++ src/app/groups/groups.service.ts
@@ -1,22 +1,21 @@
 /** Angular Imports */
-import { Injectable } from '@angular/core';
-import { HttpClient, HttpParams } from '@angular/common/http';
+import { Injectable } from "@angular/core";
+import { HttpClient, HttpParams } from "@angular/common/http";
 
 /** rxjs Imports */
-import { Observable } from 'rxjs';
+import { Observable } from "rxjs";
 
 /**
  * Groups service.
  */
 @Injectable({
-    providedIn: 'root'
+  providedIn: "root",
 })
 export class GroupsService {
-
   /**
    * @param {HttpClient} http Http Client to send requests.
    */
-  constructor(private http: HttpClient) { }
+  constructor(private http: HttpClient) {}
 
   /**
    * @param {any} filterBy Properties by which entries should be filtered.
@@ -28,18 +27,42 @@ export class GroupsService {
    */
   getGroups(filterBy: any, orderBy: string, sortOrder: string, offset?: number, limit?: number): Observable<any> {
     let httpParams = new HttpParams()
-      .set('offset', offset.toString())
-      .set('limit', limit.toString())
-      .set('sortOrder', sortOrder)
-      .set('orderBy', orderBy)
-      .set('paged', 'true');
+      .set("offset", offset.toString())
+      .set("limit", limit.toString())
+      .set("sortOrder", sortOrder)
+      .set("orderBy", orderBy)
+      .set("paged", "true");
     // filterBy: name
     filterBy.forEach(function (filter: any) {
       if (filter.value) {
         httpParams = httpParams.set(filter.type, filter.value);
       }
     });
-    return this.http.get('/groups', { params: httpParams });
+    return this.http.get("/groups", { params: httpParams });
+  }
+
+  getGroupsByCountryId(
+    filterBy: any,
+    orderBy: string,
+    sortOrder: string,
+    offset?: number,
+    limit?: number,
+    countryId?: string
+  ): Observable<any> {
+    let httpParams = new HttpParams()
+      .set("offset", offset.toString())
+      .set("limit", limit.toString())
+      .set("countryId", countryId)
+      .set("sortOrder", sortOrder)
+      .set("orderBy", orderBy)
+      .set("paged", "true");
+    // filterBy: name
+    filterBy.forEach(function (filter: any) {
+      if (filter.value) {
+        httpParams = httpParams.set(filter.type, filter.value);
+      }
+    });
+    return this.http.get("/groups", { params: httpParams });
   }
 
   /**
@@ -50,16 +73,19 @@ export class GroupsService {
    * @param {any} orphansOnly? Orphans Only.
    * @returns {Observable<any>} Groups.
    */
-  getFilteredGroups(orderBy: string, sortOrder: string, name: string, officeId?: any, orphansOnly?: any): Observable<any> {
-    let httpParams = new HttpParams()
-      .set('name', name)
-      .set('sortOrder', sortOrder)
-      .set('orderBy', orderBy);
+  getFilteredGroups(
+    orderBy: string,
+    sortOrder: string,
+    name: string,
+    officeId?: any,
+    orphansOnly?: any
+  ): Observable<any> {
+    let httpParams = new HttpParams().set("name", name).set("sortOrder", sortOrder).set("orderBy", orderBy);
     if (officeId) {
-      httpParams = httpParams.set('officeId', officeId);
+      httpParams = httpParams.set("officeId", officeId);
     }
-    httpParams = orphansOnly ? httpParams.set('orphansOnly', orphansOnly) : httpParams;
-    return this.http.get('/groups', { params: httpParams });
+    httpParams = orphansOnly ? httpParams.set("orphansOnly", orphansOnly) : httpParams;
+    return this.http.get("/groups", { params: httpParams });
   }
 
   /**
@@ -67,8 +93,8 @@ export class GroupsService {
    * @returns {Observable<any>}
    */
   getGroupsByOfficeId(officeId: number): Observable<any> {
-    const httpParams = new HttpParams().set('officeId', officeId.toString());
-    return this.http.get('/groups', { params: httpParams } );
+    const httpParams = new HttpParams().set("officeId", officeId.toString());
+    return this.http.get("/groups", { params: httpParams });
   }
 
   /**
@@ -77,8 +103,8 @@ export class GroupsService {
    * @returns {Observable<any>} Group data.
    */
   getGroupData(groupId: string, template?: string): Observable<any> {
-    let httpParams = new HttpParams().set('associations', 'all');
-    httpParams = template ? httpParams.set('template', template) : httpParams;
+    let httpParams = new HttpParams().set("associations", "all");
+    httpParams = template ? httpParams.set("template", template) : httpParams;
     return this.http.get(`/groups/${groupId}`, { params: httpParams });
   }
 
@@ -87,8 +113,7 @@ export class GroupsService {
    * @returns {Observable<any>} Group Summary data.
    */
   getGroupSummary(groupId: string): Observable<any> {
-    const httpParams = new HttpParams().set('R_groupId', groupId)
-      .set('genericResultSet', 'false');
+    const httpParams = new HttpParams().set("R_groupId", groupId).set("genericResultSet", "false");
     return this.http.get(`/runreports/GroupSummaryCounts`, { params: httpParams });
   }
 
@@ -140,7 +165,7 @@ export class GroupsService {
    * @returns {Observable<any>}
    */
   getGroupDatatables(): Observable<any> {
-    const httpParams = new HttpParams().set('apptable', 'm_group');
+    const httpParams = new HttpParams().set("apptable", "m_group");
     return this.http.get(`/datatables`, { params: httpParams });
   }
 
@@ -150,7 +175,7 @@ export class GroupsService {
    * @returns {Observable<any>}
    */
   getGroupDatatable(groupId: string, datatableName: string): Observable<any> {
-    const httpParams = new HttpParams().set('genericResultSet', 'true');
+    const httpParams = new HttpParams().set("genericResultSet", "true");
     return this.http.get(`/datatables/${datatableName}/${groupId}`, { params: httpParams });
   }
 
@@ -161,7 +186,7 @@ export class GroupsService {
    * @returns {Observable<any>}
    */
   addGroupDatatableEntry(groupId: string, datatableName: string, data: any): Observable<any> {
-    const httpParams = new HttpParams().set('genericResultSet', 'true');
+    const httpParams = new HttpParams().set("genericResultSet", "true");
     return this.http.post(`/datatables/${datatableName}/${groupId}`, data, { params: httpParams });
   }
 
@@ -172,7 +197,7 @@ export class GroupsService {
    * @returns {Observable<any>}
    */
   editGroupDatatableEntry(groupId: string, datatableName: string, data: any): Observable<any> {
-    const httpParams = new HttpParams().set('genericResultSet', 'true');
+    const httpParams = new HttpParams().set("genericResultSet", "true");
     return this.http.put(`/datatables/${datatableName}/${groupId}`, data, { params: httpParams });
   }
 
@@ -182,7 +207,7 @@ export class GroupsService {
    * @returns {Observable<any>}
    */
   deleteDatatableContent(groupId: string, datatableName: string): Observable<any> {
-    const httpParams = new HttpParams().set('genericResultSet', 'true');
+    const httpParams = new HttpParams().set("genericResultSet", "true");
     return this.http.delete(`/datatables/${datatableName}/${groupId}`, { params: httpParams });
   }
 
@@ -191,7 +216,7 @@ export class GroupsService {
    * @returns {Observable<any>} Group Command Template
    */
   getGroupCommandTemplate(command: string): Observable<any> {
-    const httpParams = new HttpParams().set('command', command);
+    const httpParams = new HttpParams().set("command", command);
     return this.http.get(`/groups/template`, { params: httpParams });
   }
 
@@ -202,7 +227,7 @@ export class GroupsService {
    * @returns {Observable<any>}
    */
   executeGroupCommand(groupId: string, command: string, data: any): Observable<any> {
-    const httpParams = new HttpParams().set('command', command);
+    const httpParams = new HttpParams().set("command", command);
     return this.http.post(`/groups/${groupId}`, data, { params: httpParams });
   }
 
@@ -212,9 +237,7 @@ export class GroupsService {
    * @returns {Observable<any>}
    */
   unAssignRoleCommand(groupId: string, roleId: any): Observable<any> {
-    const httpParams = new HttpParams()
-        .set('command', 'unassignRole')
-        .set('roleId', roleId);
+    const httpParams = new HttpParams().set("command", "unassignRole").set("roleId", roleId);
     return this.http.post(`/groups/${groupId}`, {}, { params: httpParams });
   }
 
@@ -223,7 +246,7 @@ export class GroupsService {
    * @returns {Observable<any>}
    */
   createGroup(group: any): Observable<any> {
-    return this.http.post('/groups', group);
+    return this.http.post("/groups", group);
   }
 
   /**
@@ -257,7 +280,7 @@ export class GroupsService {
    * @returns {Observable<any>}
    */
   getGroupCalendarAndTemplate(groupId: any, calendarId: any): Observable<any> {
-    const httpParams = new HttpParams().set('template', 'true');
+    const httpParams = new HttpParams().set("template", "true");
     return this.http.get(`/groups/${groupId}/calendars/${calendarId}`, { params: httpParams });
   }
 
@@ -286,7 +309,7 @@ export class GroupsService {
    * @returns {Observable<any>}
    */
   getMeetingsTemplate(groupId: any, calendarId: any): Observable<any> {
-    const httpParams = new HttpParams().set('calenderId', calendarId);
+    const httpParams = new HttpParams().set("calenderId", calendarId);
     return this.http.get(`/groups/${groupId}/meetings/template`, { params: httpParams });
   }
 
@@ -297,7 +320,7 @@ export class GroupsService {
    * @returns {Observable<any>}
    */
   assignGroupAttendance(groupId: any, calendarId: any, data: any): Observable<any> {
-    const httpParams = new HttpParams().set('calenderId', calendarId);
+    const httpParams = new HttpParams().set("calenderId", calendarId);
     return this.http.post(`/groups/${groupId}/meetings`, data, { params: httpParams });
   }
 
@@ -306,10 +329,7 @@ export class GroupsService {
    * @returns {Observable<any>} Staff Data for group.
    */
   getStaff(id: number): Observable<any> {
-    const httpParams = new HttpParams()
-        .set('officeId', id.toString())
-        .set('staffInSelectedOfficeOnly', 'true');
-    return this.http.get('/groups/template', { params: httpParams });
+    const httpParams = new HttpParams().set("officeId", id.toString()).set("staffInSelectedOfficeOnly", "true");
+    return this.http.get("/groups/template", { params: httpParams });
   }
-
 }