+++ RELEASE_NOTES.md
@@ -11,7 +11,7 @@ Release Notes
     * [SER-1538] - Implement healthy path upload in Fineract UI v2
     * [SER-1542] - Fix new loan application error
     * [SER-1617] - Fix error when accessing group member details from Group details page
-    * [SER-1536] - Implement Loan Top-up in Fineract UI 
+    * [SER-1536] - Implement Loan Top-up in Fineract UI
     * [SER-1618] - Search Client on the client list interface by multiple fields as done in the old version
     * [SER-1895] - Fix Fineract UI Signout Redirect Keycloak Error
     * [SER-1532] - Hide fields in the loan creation process
@@ -27,15 +27,18 @@ Release Notes
     * [SER-1562] - Refresh automatically after uploading a file
     * [SER-1521] - As a country admin, I should be able to add an extension charge
     * [SER-2030] - Implement loan terms and conditions
-    * [SER-2045] - Add delete button for loan product 
+    * [SER-2045] - Add delete button for loan product
     * [SER-2093] - Add qualification rules on loan products
     * [SER-2438] - Fix view client details
-    * [SER-2093] - Add qualification rules on loan products    
+    * [SER-2093] - Add qualification rules on loan products
     * [SER-2396] - Deploy fineract UI v2 on the integration cluster (Added missing envs)
     * [SER-2440] - Revise the country selection option to check for the Head Office instead of roles
     * [SER-2446] - Add ng-select as drop down in client and group creation
     * [SER-2458] - Replace mat-select-filter in all interfaces with ng select
     * [SER-2304] - Upgrade to angular 14
+    * [SER-2036] - Return groups based on the selected country
+    * [SER-2034] - Return clients based on the selected country.
+    * [SER-2039] - Return loan products based on the selected country.
     * [SER-2510] - show office hierrachy path name, show only required menus.
 
 ## Version 1.0.0 - for use with Fineract Web App